% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_bot.R
\name{format_env_vars}
\alias{format_env_vars}
\title{Format the Environment Variables List for Github Action YAML File}
\usage{
format_env_vars(env_vars, pre_line)
}
\arguments{
\item{env_vars}{A named list with the name and the value of each environment variable to use.}

\item{pre_line}{A character with the string that should be prepended to each line.}
}
\description{
Format the Environment Variables List for Github Action YAML File
}
