% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_bot.R
\name{format_r_code}
\alias{format_r_code}
\title{Format the R Code for Github Action YAML File}
\usage{
format_r_code(r_code, pre_line)
}
\arguments{
\item{r_code}{A character with the R code to be executed by the bot.}

\item{pre_line}{A character with the string that should be prepended to each line.}
}
\description{
Format the R Code for Github Action YAML File
}
