% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_bot.R
\name{make_bot}
\alias{make_bot}
\title{Configure the Bot}
\usage{
make_bot(
  r_code,
  env_vars = list(),
  bot_timeout = 15,
  repo = NULL,
  github_pat = Sys.getenv("GITHUB_PAT")
)
}
\arguments{
\item{r_code}{A character with the R code to be executed by the bot.}

\item{env_vars}{A named list with the name and the value of each environment variable to use.
E.g., `list(TWITTER_KEY = "klj3l21kj3l21k", TWITTER_SECRET = "4lh324hl34l23j4hl23j4l23j")`.}

\item{bot_timeout}{A numeric with the maximum number of minutes to let the bot script run before
GitHub automatically cancels it.}

\item{repo}{A character with the address of the Github repository in the format "username/repo".
This field is not mandatory, it will be used to try to push `env_vars` to the repository.
I.e., If the repository is "https://github.com/jcrodriguez1989/botmaker", then repo is
"jcrodriguez1989/botmaker".}

\item{github_pat}{A character with your Github personal access token (PAT). This field is not
mandatory, it will be used to try to push `env_vars` to the repository. To get your PAT, visit
"https://github.com/settings/tokens".}
}
\description{
Configure the settings to get the bot running on Github Actions.
}
