% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_bot.R
\name{set_gh_secrets}
\alias{set_gh_secrets}
\title{Write Secrets to Github}
\usage{
set_gh_secrets(secrets_list, repo, github_pat)
}
\arguments{
\item{secrets_list}{A named list with the name and the value of each environment variable to
push to Github.}

\item{repo}{A character with the address of the Github repository in the format "username/repo".}

\item{github_pat}{A character with your Github personal access token (PAT).}
}
\description{
Tries to write secrets to a Github repository.
}
